Refactor FriendlyUserName component to support also colony addresses/names
