+++ src/modules/core/components/FriendlyName/FriendlyName.css
