+++ src/modules/core/components/FriendlyName/FriendlyNamecss.d.ts
@@ -1,0 +1,1 @@
+export const main: string;
