--- src/modules/core/components/FriendlyUserName/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FriendlyUserName';
