+++ src/modules/dashboard/components/ActionsPage/ActionsPage.tsx
@@ -6,7 +6,7 @@ import { ColonyRole } from '@colony/colony-js';
 import Heading from '~core/Heading';
 import Button from '~core/Button';
 import Numeral from '~core/Numeral';
-import FriendlyUserName from '~core/FriendlyUserName';
+import FriendlyName from '~core/FriendlyName';
 import LoadingTemplate from '~pages/LoadingTemplate';
 import ActionsPageFeed, {
   ActionsPageFeedItem,
@@ -272,7 +272,7 @@ const ActionsPage = () => {
     actionType,
     initiator: (
       <span className={styles.titleDecoration}>
-        <FriendlyUserName
+        <FriendlyName
           user={initiatorProfileWithFallback}
           autoShrinkAddress
         />
@@ -280,7 +280,7 @@ const ActionsPage = () => {
     ),
     recipient: (
       <span className={styles.titleDecoration}>
-        <FriendlyUserName
+        <FriendlyName
           user={recipientProfileWithFallback}
           autoShrinkAddress
         />
